refactor(features-comparison, pricing-tiers): improve data structures and enhance type definitions
