+++ apps/demo/src/app/data/featuresComparisonData.ts
@@ -1,22 +1,25 @@
 import { featuresList } from './featuresData';
 
-export interface PricingTier {
+export interface FeatureTier {
   name: string;
   features: string[];
 }
 
+const FREE_TIER_FEATURES_COUNT = 3;
+const PRO_TIER_FEATURES_COUNT = 4;
+
 /**
- * An array of pricing tiers, each with a name and a list of features.
+ * An array of feature tiers, each with a name and a list of features.
  * The features are subsets of the featuresList.
  */
-export const pricingTiers: PricingTier[] = [
+export const featureTiers: FeatureTier[] = [
   {
     name: 'Free',
-    features: featuresList.slice(0, 3),
+    features: featuresList.slice(0, FREE_TIER_FEATURES_COUNT),
   },
   {
     name: 'Pro',
-    features: featuresList.slice(0, 4),
+    features: featuresList.slice(0, PRO_TIER_FEATURES_COUNT),
   },
   {
     name: 'Enterprise',
@@ -25,13 +28,13 @@ export const pricingTiers: PricingTier[] = [
 ];
 
 /**
- * An array of objects that compare features across different pricing tiers.
+ * An array of objects that compare features across different feature tiers.
  * Each object contains a feature and boolean values indicating its availability
  * in the basic, pro, and enterprise tiers.
  */
 export const featuresComparisonData = featuresList.map((feature) => ({
   feature,
-  basic: featuresList.slice(0, 3).includes(feature),
-  pro: featuresList.slice(0, 4).includes(feature),
+  basic: featuresList.slice(0, FREE_TIER_FEATURES_COUNT).includes(feature),
+  pro: featuresList.slice(0, PRO_TIER_FEATURES_COUNT).includes(feature),
   enterprise: featuresList.includes(feature),
 }));
